+++ demo/src/main/java/raven/modal/demo/Demo.java
@@ -13,6 +13,8 @@ import java.awt.*;
 
 public class Demo extends JFrame {
 
+    public static final String DEMO_VERSION = "1.2.1";
+
     public Demo() {
         init();
     }