+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fixed component not install when push modal
 - Updated select drawer menu when undo and redo form
 - Fixed component orientation `RTL` not update when show form
+- Add `about` information
 
 ## [1.2.0] - 2024-09-21